Add support for custom-parsing single references
The url_resolver config option was only working on references
being part of toctrees (why?). This applies the url_resolver
on all references so one can do interesting things like
`[text](github:object.py)` and have that be expanded to a real url
pointing to the file in the repo online.